[#13825] refactor: extract student and instructor search tables into separate components - Created AdminStudentSearchTableComponent following AdminAccountSearchTableComponent pattern - Created AdminInstructorSearchTableComponent following AdminAccountSearchTableComponent pattern - Moved AdminAccountSearchTableComponent from components/admin-search-tables into admin-search-page - Deleted admin-search-tables folder and account-search-table-model.ts - Applied ngOnChanges pattern consistently across all 3 table components - Moved tests from parent spec to respective child spec files - Updated parent component to use new component selectors
