--- src/web/app/components/admin-search-tables/account-search-table-model.ts
@@ -1,1 +1,0 @@
-export type { AccountRequestTableRowModel } from '../account-requests-table/account-request-table-model';
