+++ src/web/app/pages-admin/admin-search-page/admin-account-search-table/admin-account-search-table.component.scss
