+++ src/web/app/pages-admin/admin-search-page/admin-search-page.component.ts
@@ -1,31 +1,20 @@
-import { NgClass, KeyValuePipe } from '@angular/common';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { NgbCollapse, NgbModalRef, NgbTooltip } from '@ng-bootstrap/ng-bootstrap';
 import { finalize } from 'rxjs/operators';
-import { AccountService } from '../../../services/account.service';
-import { EmailGenerationService } from '../../../services/email-generation.service';
-import { InstructorService } from '../../../services/instructor.service';
 import { LoadingBarService } from '../../../services/loading-bar.service';
 import {
   AccountRequestSearchResult,
   AdminSearchResult,
-  FeedbackSessionsGroup,
   InstructorAccountSearchResult,
   SearchService,
   StudentAccountSearchResult,
 } from '../../../services/search.service';
-import { SimpleModalService } from '../../../services/simple-modal.service';
 import { StatusMessageService } from '../../../services/status-message.service';
-import { StudentService } from '../../../services/student.service';
 import { ApiConst } from '../../../types/api-const';
-import { Email, RegenerateKey } from '../../../types/api-output';
-import { AccountRequestTableRowModel } from '../../components/account-requests-table/account-request-table-model';
-import { AdminAccountSearchTableComponent } from '../../components/admin-search-tables/admin-account-search-table.component';
-import { AjaxLoadingComponent } from '../../components/ajax-loading/ajax-loading.component';
-import { SimpleModalType } from '../../components/simple-modal/simple-modal-type';
+import { AdminAccountSearchTableComponent } from './admin-account-search-table/admin-account-search-table.component';
+import { AdminInstructorSearchTableComponent } from './admin-instructor-search-table/admin-instructor-search-table.component';
+import { AdminStudentSearchTableComponent } from './admin-student-search-table/admin-student-search-table.component';
 import { ErrorMessageOutput } from '../../error-message-output';
-import { SearchTermsHighlighterPipe } from '../../pipes/search-terms-highlighter.pipe';
 
 /**
  * Admin search page.
@@ -36,13 +25,9 @@ import { SearchTermsHighlighterPipe } from '../../pipes/search-terms-highlighter
   styleUrls: ['./admin-search-page.component.scss'],
   imports: [
     FormsModule,
-    NgClass,
-    NgbTooltip,
-    AjaxLoadingComponent,
     AdminAccountSearchTableComponent,
-    KeyValuePipe,
-    SearchTermsHighlighterPipe,
-    NgbCollapse,
+    AdminInstructorSearchTableComponent,
+    AdminStudentSearchTableComponent,
   ],
 })
 export class AdminSearchPageComponent {
@@ -50,20 +35,12 @@ export class AdminSearchPageComponent {
   searchString = '';
   instructors: InstructorAccountSearchResult[] = [];
   students: StudentAccountSearchResult[] = [];
-  accountRequests: AccountRequestTableRowModel[] = [];
+  accountRequests: AccountRequestSearchResult[] = [];
   characterLimit = 100;
-
-  isRegeneratingInstructorKeys: boolean[] = [];
-  isRegeneratingStudentKeys: boolean[] = [];
 
   constructor(
     private statusMessageService: StatusMessageService,
-    private simpleModalService: SimpleModalService,
-    private accountService: AccountService,
-    private instructorService: InstructorService,
-    private studentService: StudentService,
     private searchService: SearchService,
-    private emailGenerationService: EmailGenerationService,
     private loadingBarService: LoadingBarService,
   ) {}
 
@@ -96,13 +73,7 @@ export class AdminSearchPageComponent {
           this.instructors = resp.instructors;
           this.students = resp.students;
           this.accountRequests = this.formatAccountRequests(resp.accountRequests);
-          this.hideAllInstructorsLinks();
-          this.hideAllStudentsLinks();
 
-          this.isRegeneratingInstructorKeys = new Array(this.instructors.length).fill(false);
-          this.isRegeneratingStudentKeys = new Array(this.students.length).fill(false);
-
-          // prompt user to use more specific terms if search results limit reached
           const limit: number = ApiConst.SEARCH_QUERY_SIZE_LIMIT;
           const limitsReached: string[] = [];
           if (this.students.length >= limit) {
@@ -129,267 +100,20 @@ export class AdminSearchPageComponent {
       });
   }
 
-  private formatAccountRequests(accountRequests: AccountRequestSearchResult[]): AccountRequestTableRowModel[] {
-    return accountRequests.map((accountRequest: AccountRequestSearchResult): AccountRequestTableRowModel => {
+  private formatAccountRequests(accountRequests: AccountRequestSearchResult[]): AccountRequestSearchResult[] {
+    return accountRequests.map((accountRequest: AccountRequestSearchResult): AccountRequestSearchResult => {
       return {
         id: accountRequest.id,
         name: accountRequest.name,
         email: accountRequest.email,
         status: accountRequest.status,
-        instituteAndCountry: accountRequest.institute,
+        institute: accountRequest.institute,
         createdAtText: accountRequest.createdAtText,
         registeredAtText: accountRequest.registeredAtText || '',
         comments: accountRequest.comments,
         registrationLink: accountRequest.registrationLink,
         showLinks: accountRequest.showLinks,
       };
-    });
-  }
-
-  /**
-   * Shows all instructors' links in the page.
-   */
-  showAllInstructorsLinks(): void {
-    for (const instructor of this.instructors) {
-      instructor.showLinks = true;
-    }
-  }
-
-  /**
-   * Hides all instructors' links in the page.
-   */
-  hideAllInstructorsLinks(): void {
-    for (const instructor of this.instructors) {
-      instructor.showLinks = false;
-    }
-  }
-
-  /**
-   * Shows all students' links in the page.
-   */
-  showAllStudentsLinks(): void {
-    for (const student of this.students) {
-      student.showLinks = true;
-    }
-  }
-
-  /**
-   * Hides all students' links in the page.
-   */
-  hideAllStudentsLinks(): void {
-    for (const student of this.students) {
-      student.showLinks = false;
-    }
-  }
-
-  /**
-   * Resets the instructor's Google ID.
-   */
-  resetInstructorGoogleId(instructor: InstructorAccountSearchResult, event: any): void {
-    if (event) {
-      event.preventDefault();
-      event.stopPropagation();
-    }
-
-    const modalContent = `Are you sure you want to reset the Google account ID currently associated for
-        <strong>${instructor.name}</strong> in the course <strong>${instructor.courseId}</strong>?
-        The user will need to re-associate their account with a new Google ID.`;
-    const modalRef: NgbModalRef = this.simpleModalService.openConfirmationModal(
-      `Reset <strong>${instructor.name}</strong>'s Google ID?`,
-      SimpleModalType.WARNING,
-      modalContent,
-    );
-
-    modalRef.result.then(
-      () => {
-        this.accountService.resetInstructorAccount(instructor.courseId, instructor.email).subscribe({
-          next: () => {
-            this.search();
-            this.statusMessageService.showSuccessToast("The instructor's Google ID has been reset.");
-          },
-          error: (resp: ErrorMessageOutput) => {
-            this.statusMessageService.showErrorToast(resp.error.message);
-          },
-        });
-      },
-      () => {},
-    );
-  }
-
-  /**
-   * Resets the student's Google ID.
-   */
-  resetStudentGoogleId(student: StudentAccountSearchResult, event: any): void {
-    if (event) {
-      event.preventDefault();
-      event.stopPropagation();
-    }
-    const modalContent = `Are you sure you want to reset the Google account ID currently associated for
-        <strong>${student.name}</strong> in the course <strong>${student.courseId}</strong>?
-        The user will need to re-associate their account with a new Google ID.`;
-    const modalRef: NgbModalRef = this.simpleModalService.openConfirmationModal(
-      `Reset <strong>${student.name}</strong>'s Google ID?`,
-      SimpleModalType.WARNING,
-      modalContent,
-    );
-
-    modalRef.result.then(
-      () => {
-        this.accountService.resetStudentAccount(student.courseId, student.email).subscribe({
-          next: () => {
-            student.googleId = '';
-            this.statusMessageService.showSuccessToast("The student's Google ID has been reset.");
-          },
-          error: (resp: ErrorMessageOutput) => {
-            this.statusMessageService.showErrorToast(resp.error.message);
-          },
-        });
-      },
-      () => {},
-    );
-  }
-
-  /**
-   * Regenerates the student's registration key.
-   */
-  regenerateStudentKey(student: StudentAccountSearchResult, index: number): void {
-    this.isRegeneratingStudentKeys[index] = true;
-    const modalContent = `Are you sure you want to regenerate the registration key for
-        <strong>${student.name}</strong> for the course <strong>${student.courseId}</strong>?
-        An email will be sent to the student with all the new course registration and feedback session links.`;
-    const modalRef: NgbModalRef = this.simpleModalService.openConfirmationModal(
-      `Regenerate <strong>${student.name}</strong>'s course links?`,
-      SimpleModalType.WARNING,
-      modalContent,
-    );
-
-    modalRef.dismissed.subscribe(() => {
-      this.isRegeneratingStudentKeys[index] = false;
-    });
-
-    modalRef.result.then(
-      () => {
-        this.studentService.regenerateStudentKey(student.courseId, student.email).subscribe({
-          next: (resp: RegenerateKey) => {
-            this.statusMessageService.showSuccessToast(resp.message);
-            this.updateDisplayedStudentCourseLinks(student, resp.newRegistrationKey);
-            this.isRegeneratingStudentKeys[index] = false;
-          },
-          error: (response: ErrorMessageOutput) => {
-            this.statusMessageService.showErrorToast(response.error.message);
-            this.isRegeneratingStudentKeys[index] = false;
-          },
-        });
-      },
-      () => {},
-    );
-  }
-
-  /**
-   * Regenerates the instructor's registration key.
-   */
-  regenerateInstructorKey(instructor: InstructorAccountSearchResult, index: number): void {
-    this.isRegeneratingInstructorKeys[index] = true;
-    const modalContent = `Are you sure you want to regenerate the registration key for
-        <strong>${instructor.name}</strong> for the course <strong>${instructor.courseId}</strong>?
-        An email will be sent to the instructor with all the new course registration and feedback session links.`;
-    const modalRef: NgbModalRef = this.simpleModalService.openConfirmationModal(
-      `Regenerate <strong>${instructor.name}</strong>'s course links?`,
-      SimpleModalType.WARNING,
-      modalContent,
-    );
-
-    modalRef.dismissed.subscribe(() => {
-      this.isRegeneratingInstructorKeys[index] = false;
-    });
-
-    modalRef.result.then(
-      () => {
-        this.instructorService.regenerateInstructorKey(instructor.courseId, instructor.email).subscribe({
-          next: (resp: RegenerateKey) => {
-            this.statusMessageService.showSuccessToast(resp.message);
-            this.updateDisplayedInstructorCourseLinks(instructor, resp.newRegistrationKey);
-            this.isRegeneratingInstructorKeys[index] = false;
-          },
-          error: (response: ErrorMessageOutput) => {
-            this.statusMessageService.showErrorToast(response.error.message);
-            this.isRegeneratingInstructorKeys[index] = false;
-          },
-        });
-      },
-      () => {},
-    );
-  }
-
-  /**
-   * Updates the student's displayed course join and feedback session links with the value of the newKey.
-   */
-  private updateDisplayedStudentCourseLinks(student: StudentAccountSearchResult, newKey: string): void {
-    const updateSessions = (sessions: FeedbackSessionsGroup): void => {
-      Object.keys(sessions).forEach((key: string): void => {
-        sessions[key].feedbackSessionUrl = this.getUpdatedUrl(sessions[key].feedbackSessionUrl, newKey);
-      });
-    };
-
-    student.courseJoinLink = this.getUpdatedUrl(student.courseJoinLink, newKey);
-    updateSessions(student.awaitingSessions);
-    updateSessions(student.openSessions);
-    updateSessions(student.notOpenSessions);
-    updateSessions(student.publishedSessions);
-  }
-
-  /**
-   * Updates the instructor's displayed course join and feedback session links with the value of the newKey.
-   */
-  private updateDisplayedInstructorCourseLinks(instructor: InstructorAccountSearchResult, newKey: string): void {
-    const updateSessions = (sessions: FeedbackSessionsGroup): void => {
-      Object.keys(sessions).forEach((key: string): void => {
-        sessions[key].feedbackSessionUrl = this.getUpdatedUrl(sessions[key].feedbackSessionUrl, newKey);
-      });
-    };
-
-    instructor.courseJoinLink = this.getUpdatedUrl(instructor.courseJoinLink, newKey);
-    updateSessions(instructor.awaitingSessions);
-    updateSessions(instructor.openSessions);
-    updateSessions(instructor.notOpenSessions);
-    updateSessions(instructor.publishedSessions);
-  }
-
-  /**
-   * Returns the URL after replacing the value of the `key` parameter with that of the new key.
-   */
-  private getUpdatedUrl(link: string, newVal: string): string {
-    const param = 'key';
-    const regex = new RegExp(`(${param}=)[^&]+`);
-
-    return link.replace(regex, `$1${newVal}`);
-  }
-
-  /**
-   * Open up an email populated with content for course join invitation.
-   */
-  openCourseJoinEmail(studentId: string): void {
-    this.emailGenerationService.getCourseJoinEmail(studentId).subscribe({
-      next: (email: Email) => {
-        window.location.href = `mailto:${email.recipient}` + `?Subject=${email.subject}` + `&body=${email.content}`;
-      },
-      error: (err: ErrorMessageOutput) => {
-        this.statusMessageService.showErrorToast(err.error.message);
-      },
-    });
-  }
-
-  /**
-   * Open up an email populated with content for feedback session reminder.
-   */
-  openFeedbackSessionReminderEmail(studentId: string, fsId: string): void {
-    this.emailGenerationService.getFeedbackSessionReminderEmail(studentId, fsId).subscribe({
-      next: (email: Email) => {
-        window.location.href = `mailto:${email.recipient}` + `?Subject=${email.subject}` + `&body=${email.content}`;
-      },
-      error: (err: ErrorMessageOutput) => {
-        this.statusMessageService.showErrorToast(err.error.message);
-      },
     });
   }
 }
